refactor: 회원, 경매 서비스 로직 제
