--- src/main/java/org/indoles/receiptserviceserver/core/auction/controller/CurrentTime.java
@@ -1,11 +1,0 @@
-package org.indoles.receiptserviceserver.core.auction.controller;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(value = ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface CurrentTime {
-}
