--- src/main/java/org/indoles/receiptserviceserver/core/auction/controller/CurrentTimeArgumentResolver.java
@@ -1,25 +1,0 @@
-package org.indoles.receiptserviceserver.core.auction.controller;
-
-import org.springframework.core.MethodParameter;
-import org.springframework.stereotype.Component;
-import org.springframework.web.bind.support.WebDataBinderFactory;
-import org.springframework.web.context.request.NativeWebRequest;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.method.support.ModelAndViewContainer;
-
-import java.time.LocalDateTime;
-
-@Component
-public class CurrentTimeArgumentResolver implements HandlerMethodArgumentResolver {
-
-    @Override
-    public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.hasParameterAnnotation(CurrentTime.class);
-    }
-
-    @Override
-    public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer,
-                                  NativeWebRequest webRequest, WebDataBinderFactory binderFactory) throws Exception {
-        return LocalDateTime.now();
-    }
-}
