--- src/main/java/org/indoles/receiptserviceserver/core/auction/controller/dto/PurchaseResponse.java
@@ -1,8 +1,0 @@
-package org.indoles.receiptserviceserver.core.auction.controller.dto;
-
-import java.util.UUID;
-
-public record PurchaseResponse(
-        UUID receiptId
-) {
-}
