--- src/main/java/org/indoles/receiptserviceserver/core/auction/domain/AuctionStatus.java
@@ -1,21 +1,0 @@
-package org.indoles.receiptserviceserver.core.auction.domain;
-
-public enum AuctionStatus {
-    WAITING, RUNNING, SOLD_OUT, FINISHED;
-
-    public boolean isWaiting() {
-        return this == WAITING;
-    }
-
-    public boolean isRunning() {
-        return this == RUNNING;
-    }
-
-    public boolean isSoldOut() {
-        return this == SOLD_OUT;
-    }
-
-    public boolean isFinished() {
-        return this == FINISHED;
-    }
-}
