--- src/main/java/org/indoles/receiptserviceserver/core/auction/domain/PricePolicy.java
@@ -1,30 +1,0 @@
-package org.indoles.receiptserviceserver.core.auction.domain;
-
-import com.fasterxml.jackson.annotation.JsonSubTypes;
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
-import com.fasterxml.jackson.annotation.JsonTypeInfo.As;
-import com.fasterxml.jackson.annotation.JsonTypeInfo.Id;
-
-@JsonTypeInfo(
-        use = Id.NAME,
-        include = As.EXISTING_PROPERTY,
-        property = "type"
-)
-@JsonSubTypes({
-        @JsonSubTypes.Type(value = PercentagePricePolicy.class, name = "PERCENTAGE"),
-        @JsonSubTypes.Type(value = ConstantPricePolicy.class, name = "CONSTANT")
-})
-public interface PricePolicy {
-
-    long calculatePriceAtVariation(long price, long variationCount);
-
-    static PricePolicy createPercentagePricePolicy(int discountRate) {
-        return new PercentagePricePolicy(discountRate);
-    }
-
-    static PricePolicy createConstantPricePolicy(int variationWidth) {
-        return new ConstantPricePolicy(variationWidth);
-    }
-
-    PricePolicyType getType();
-}
