--- src/main/java/org/indoles/receiptserviceserver/core/auction/domain/PricePolicyType.java
@@ -1,5 +1,0 @@
-package org.indoles.receiptserviceserver.core.auction.domain;
-
-public enum PricePolicyType {
-    PERCENTAGE, CONSTANT
-}
