--- src/main/java/org/indoles/receiptserviceserver/core/auction/infra/AuctionQueryDslRepository.java
@@ -1,17 +1,0 @@
-package org.indoles.receiptserviceserver.core.auction.infra;
-
-
-
-import org.indoles.receiptserviceserver.core.auction.dto.AuctionSearchCondition;
-import org.indoles.receiptserviceserver.core.auction.dto.SellerAuctionSearchCondition;
-import org.indoles.receiptserviceserver.core.auction.entity.AuctionEntity;
-
-import java.util.List;
-
-public interface AuctionQueryDslRepository {
-
-    List<AuctionEntity> findAllBy(AuctionSearchCondition condition);
-
-    List<AuctionEntity> findAllBy(SellerAuctionSearchCondition condition);
-
-}
