--- src/main/java/org/indoles/receiptserviceserver/core/member/controller/AuthenticationArgumentResolver.java
@@ -1,28 +1,0 @@
-package org.indoles.receiptserviceserver.core.member.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.indoles.receiptserviceserver.core.member.domain.AuthenticationContext;
-import org.springframework.core.MethodParameter;
-import org.springframework.stereotype.Component;
-import org.springframework.web.bind.support.WebDataBinderFactory;
-import org.springframework.web.context.request.NativeWebRequest;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.method.support.ModelAndViewContainer;
-
-@Component
-@RequiredArgsConstructor
-public class AuthenticationArgumentResolver implements HandlerMethodArgumentResolver {
-
-    private final AuthenticationContext authenticationContext;
-
-    @Override
-    public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.hasParameterAnnotation(Login.class);
-    }
-
-    @Override
-    public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer,
-                                  NativeWebRequest webRequest, WebDataBinderFactory binderFactory) throws Exception {
-        return authenticationContext.getPrincipal();
-    }
-}
