--- src/main/java/org/indoles/receiptserviceserver/core/member/controller/Login.java
@@ -1,11 +1,0 @@
-package org.indoles.receiptserviceserver.core.member.controller;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(value = RetentionPolicy.RUNTIME)
-@Target(value = {ElementType.PARAMETER})
-public @interface Login {
-}
