--- src/main/java/org/indoles/receiptserviceserver/core/member/controller/Roles.java
@@ -1,16 +1,0 @@
-package org.indoles.receiptserviceserver.core.member.controller;
-
-
-import org.indoles.receiptserviceserver.core.member.domain.Role;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target({ElementType.METHOD})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Roles {
-
-    Role[] value();
-}
