--- src/main/java/org/indoles/receiptserviceserver/core/member/controller/dto/SignInRequestInfo.java
@@ -1,27 +1,0 @@
-package org.indoles.receiptserviceserver.core.member.controller.dto;
-
-
-
-import org.indoles.receiptserviceserver.global.exception.BadRequestException;
-import org.indoles.receiptserviceserver.global.exception.ErrorCode;
-
-import java.util.Objects;
-
-public record SignInRequestInfo(
-        String signInId,
-        String password
-) {
-
-    private static final String ERROR_NULL_VALUE = "%s는 Null일 수 없습니다.";
-
-    public SignInRequestInfo {
-        validateNotNull(signInId, "로그인 ID");
-        validateNotNull(password, "로그인 패스워드");
-    }
-
-    private void validateNotNull(Object value, String fieldName) {
-        if (Objects.isNull(value)) {
-            throw new BadRequestException(String.format(ERROR_NULL_VALUE, fieldName), ErrorCode.G000);
-        }
-    }
-}
