--- src/main/java/org/indoles/receiptserviceserver/core/member/controller/dto/SignInResponseInfo.java
@@ -1,9 +1,0 @@
-package org.indoles.receiptserviceserver.core.member.controller.dto;
-
-
-import org.indoles.receiptserviceserver.core.member.domain.Role;
-
-public record SignInResponseInfo(
-        Role role
-) {
-}
