--- src/main/java/org/indoles/receiptserviceserver/core/member/domain/MemberRepository.java
@@ -1,14 +1,0 @@
-package org.indoles.receiptserviceserver.core.member.domain;
-
-import java.util.Optional;
-
-public interface MemberRepository {
-
-    boolean isExist(String signInId);
-
-    Member save(Member member);
-
-    Optional<Member> findById(Long id);
-
-    Optional<Member> findBySignInId(String signInId);
-}
