+++ src/main/java/org/indoles/receiptserviceserver/core/receipt/domain/enums/ReceiptStatus.java
@@ -1,4 +1,4 @@
-package org.indoles.receiptserviceserver.core.receipt.domain;
+package org.indoles.receiptserviceserver.core.receipt.domain.enums;
 
 import lombok.Getter;
